+++ Back-End/src/main/java/com/example/polls/util/Helper.java
@@ -2,7 +2,7 @@ package com.example.polls.util;
 
 public class Helper {
 
-    public static String twoDigitString(int number) {
+    public static String twoDigitString(float number) {
         if (number == 0)
             return "00";
         if (number / 10 == 0)
