+++ Back-End/src/main/java/com/example/polls/util/TimeHandler.java
@@ -61,9 +61,9 @@ public class TimeHandler {
 
 
     public static String formattedTime(TimeRequest time){
-        int hours = time.getTime() / 3600;
-        int minutes = (time.getTime() % 3600) / 60;
-        int seconds = time.getTime() % 60;
+        float hours = time.getTime() / 3600;
+        float minutes = (time.getTime() % 3600) / 60;
+        float seconds = time.getTime() % 60;
         String formattedTime = "00" + ":" + Helper.twoDigitString(hours) + ":" + Helper.twoDigitString(minutes) + ":" + Helper.twoDigitString(seconds);
 
         return formattedTime;
